rename attention_multiple to basefnmt and update README
